backend/requirements.txt
fastapi
uvicorn[standard]
python-dotenv
celery
redis # For Celery broker connection
